Use macro config for entity search form
Simplify components page form examples by removing middleware
