+++ src/apps/components/form/index.js
@@ -1,4 +1,4 @@
-const constants = require('./constants')
+const constants = require('./macros')
 const controllers = require('./controllers')
 
 module.exports = {
